+++ rrule/src/core/mod.rs
@@ -8,7 +8,9 @@ pub use self::rruleset::RRuleSet;
 pub(crate) use datetime::{DateTime, Time};
 
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy)]
+/// An empty struct to keep the validated stage
 pub struct Validated;
 
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy)]
+/// An empty struct to keep the unvalidated (or not-yet-validated) stage
 pub struct Unvalidated;