+++ rrule/src/tests/mod.rs
@@ -1,0 +1,10 @@
+#![cfg(test)]
+
+mod common;
+mod datetime;
+mod edge_cases;
+mod regression;
+mod rfc_tests;
+mod rrule;
+mod rruleset;
+mod serde;