+++ rrule/src/tests/rrule.rs
@@ -1,8 +1,6 @@
-mod common;
-
+use crate::tests::common::{test_recurring_rrule, ymd_hms};
+use crate::{Frequency, NWeekday, RRule, RRuleSet, Weekday};
 use chrono::{Datelike, TimeZone};
-use common::{test_recurring_rrule, ymd_hms};
-use rrule::{Frequency, NWeekday, RRule, RRuleSet, Weekday};
 
 #[test]
 fn yearly() {